requirements.txt
python-telegram-bot[job-queue]==22.5
python-dotenv==1.1.1
APScheduler==3.10.4
